Refresh pkgver (v23)
And try to update the dependencies, based on `mkosi dependencies`
output. There seems to be very little indication what mkosi’s
dependencies mean, and at least some of them are clearly just for
developer convenience inside the built image (rather than being needed
for any step in building the image), so I’ve not added everything that
mkosi prints; but hopefully this is at least an improvement. But I
barely use mkosi myself these days, so I’m always glad to receive
comments on these dependencies on the AUR or via email :)