Update source= and url=, bump pkgver=
git:// URLs are somewhere between deprecated (according to the
out-of-date notification by a821) and non-functional (as far as I know
for GitHub), so use git+https:// instead; while we’re at it, update the
regular URL to use HTTPS too (and, apparently, www2 instead of www – the
old URL redirects to the new one).

This doesn’t actually produce a working build for me, so it’s mostly
pointless, to be honest. If anyone else is interested in still getting
dgsh to run, maybe this is slightly more helpful as a starting point –
you may also find these additional CFLAGS useful (though they’re still
not sufficient to make it work):

    make $MAKEFLAGS CFLAGS="$CFLAGS -Wno-implicit-function-declaration -Wno-implicit-int -Wno-incompatible-pointer-types"

But my general impression is that upstream dgsh development has also
stalled, and probably not much more is going to happen there.