Improve preview mode, and disable trailing blackslash for cat input mode.
The preview feature would not work correctly if the user prompt had spaces between the input proper and the trailing forward slash for activating the preview mode.
Before, as in chat completions the script removes trailing spaces, the preview user prompt would differ from the old user prompt in a second-step confirmation checking test, while trailing spaces should not be considered a new user prompt for this feature.

Also, disable the feature in which typing a trailing backslash added a new line to the user prompt and set the cat prompter mode. The idea was good at first but I rarely ever used this feature and got confused by it a few times (as I dont use it, I had forgotten about this feature), so better to disable this feature for a more simple user interface and experience.

There are many options to set the cat prompter or the text editor interfaces, so this feature now became an overload.