Perl/CPAN Module Struct::Diff
This is one of the perl-dependencies of Koha: https://koha-community.org/
Start "./koha_perl_deps.pl -r -a" after downloading and unpacking
the koha-tarball for a list of all required packages (version 24.05.00).

This is with the cpan2aur-bugfix,
see https://rt.cpan.org/Public/Bug/Display.html?id=123153 for details.