upgpkg: teleport 16.0.3-1
- Update license to AGPLv3, per announcement:
  https://goteleport.com/blog/teleport-community-license/
- Add tbot and fdpass-teleport binaries to client package
- Revise messaging about client package, as the move is no
  longer "recent"

NOTE: This package sometimes fails to compile, due to what appears to be
a race condition building ring (https://docs.rs/ring/latest/ring/). I am
(reluctantly) releasing the updated package with this condition unfixed.

NOTE 2: Past revisions of this package indicated use of the Apache 2.0
license. Because this package has always built Teleport from source, I
believe the license field was previously incorrect. To the best of my
knowledge, the correct license would have been AGPLv3 for all versions
of this package. THIS IS NOT LEGAL ADVICE! If in doubt, please speak
with your attorney.