Minor improvements: change internal function glob for session command name matching, and simplify last update code.
Be less strict in the glob test, as that had something to do with trying to work with filenames with white spaces a long time ago, and it does work for some chat commands I am just not sure if all of the commands and in what circunstances, but lets be a little more permissive when the session name end with ".tsv" extension. Maybe only I know this and use it for testing purposes only, though.

This will impact the "/copy" command for example when the last history file name ends with ".tsv", which is really highly unlikely anyway...