Add alias "/." to "/fork current", add new syntax to load custom prompt when first positional argument is ".prompt_name".
Add another command line syntax for setting a custom prompt as a command line argument instead of it being an argument to option `-S`. So, the old syntax works as `-S ..prompt`, in which the dot or double dot operators load a custom prompt. `.` differs from `..` in which the single dot operator allows for a single shot editing o the prompt, while double dots just load it silently.

The new syntax allows that the first positional argument be the operator plus name of the prompt, such as `chatgpt.sh -cc ..prompt`.

Add an alias to the chat command `!fork current` as `!.` or `/.`. This makes it simple to fork a previous session from the current/defaults history file. For example, if the user wants to load an older session from the defaults history file, the script invocation can be `chatgpt.sh -cc "/fork current"`, `chatgpt.sh -cc "/fork."`, and now simply  `chatgpt.sh -cc "/."`.

Dont exit script on custom prompt loading error, instead just clear instruction.

Update docs and rewrite examples for loading custom prompts.