govarnam-ibus-bin: upgpkg: 1.6.4-1
Govarnam-ibus is licensed under MPL-1.1, while govarnam itself is
licensed under AGPL-3.0, so use correct license. Also, upstream renamed
Kannada language identifier from ka to kn, so adapt for that, but the icon is still
named ka, a pull request is open, but until it is merged, rename said
icon file in prepare().

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>