Fix recognising last argument as text, pdf or url for the text dump feature.
Improved recognition of txt files, and urls, since when we added this feature in v0.62.7.

When the last argument of the user prompt is a text or pdf file, or a recognisable url, the script will dump and add it to the tail of the current user prompt before making a call.

A large bash history may be slow to load when executing the script at start-up!

So, we added a warning that the bash history file 'history_bash' file located in the cache directory, is too big. This file may grow very large when dumping a lot of text content to submission for the LLMs. The warning allows the user to decide if he wants to back the history file or continue using it.

Also, decrese $HISTSIZE variable from 512 to 256 and expose it to user configuration file.