Build configuration changes
- enabled hwloc
- dropped building of internal fftw (suggested by Andrey Alekseenko)

Arch-supplied fftw is built with AVX, but not AVX2 and AVX512. While
this will not provide optimal performance in some cases, it greatly
simplifies the build procedure and makes it more reliable. It is still
possible to enable building of internal FFTW by uncommenting the line
in the PKGBUILD (just like CUDA/OpenCL).