use SPDX license identifier, add some edge-case deps
libass is always required. It's usually installed as a dependency of
ffmpeg, but could be missing if e.g. ffmpeg is statically compiled.

vapoursynth (an optional dep for mpv) is currently a dependency of
ffmpeg as well, but in case that changes in the future, add a feature
option here for people who want to ensure it's enabled.