Switch to using 'bin/package install'
Most of the installation is now handled by the bin/package script
itself, rather than being done manually. For this to work the
file(1) command bin/package uses to determine the architecture must be
invoked with the --no-sandbox flag, since fakeroot will ordinarily
prevent it from functioning because of seccomp restrictions. The patch
for this was rejected upstream.
For more information: https://github.com/ksh93/ksh/pull/769