Add file picker dialog (easier to pick a file), add "/pick" and "/photo" chat commands.
Termux environment now tells the script to save output files to `$HOME/storage/downloads` (symlink to directory) instead of `$HOME/Downloads` when possible (this may be set in the config file `$OUTDIR`).

Environment and config file `$RESTART` and `$START` for their respective setting can now be set to null. This will be respected. Before, if these variables were null, the Q: and A: sequences would be added in chat mode, for example. If these are set and null, no restart or start sequences will be used in the request body of text completions.

This should be completely compatible with previous user workflows.

**New chat commands**:

The new  File Picker:

   /pick [dir]

The picker will try to use a graphical interface for file picking, if not, then defaults to vifm, nnn and ranger for file picking before falling back to `dialog`.

There is a hack-like version of a file picker for Termux Android, too!!

Should make the life of everyone easier to pick files from the terminal!

Also, if the `/pick` command is typed at the end of a prompt, the name  of the picked file will be appended to the current user prompt!

In Termux, command `/photo [index]` will take a picture with the cell phone came and append the filename to the user prompt.

We try to set the correct orientation of the recently taken photograph with `exiftran` or `imagemagick` (Termux).

Chat options -c, -cc and -C will toggle between the endpoints: text chat, native chat and text cmpls.

Improve User Interface of the session and prompt system a little.

Update man and help pages, as well readme.md file in the repo.