Restore quick and dirty work-around for host build failure.
While both clean chroot builds as well as container builds are successful without it,
multiple AUR users have reported that their systems still fail on this.
As their systems are supposedly fully updated and their builds are done in a clean chroot,
there is something making this inconsistent.

This restores the hacky work-around until the WHY can be determined.