Remove forced disablement of LTO via `{C,CXX,LD}FLAGS`
This used to be disabled in commit 2673045 ("Disable LTO") where,
according to the comment (that was accidentally removed by me in commit
2bde931 ("Cleanup and make the package build again in 2022")), linking
failed on an unknown symbol.

As this issue can no longer be reproduced, and above all stripping
`,-z,now` could lead to an invalid command line when this is the only
config on the users' `-Wl` argument (i..e `-Wl,-z,now` leaves an invalid
`-Wl` in place), remove the subtitutions to re-enable LTO again. This
was tested in an environment where the substituted flags are present
in `makepkg.conf`.