kdumpst: Bump to v1.2
This version has (finally!) a functional fix to the initcpio generation
with regards GPU drivers . It shouldn't include them anymore, but even
if some custom hook does include them, we now ban the common GFX drivers
(and DRM core init function) on the kdump command-line!

Special thanks to the bug reporters (@schardev and @flifloo) and to
Foxboron / noclaf, at the Arch linux IRC channels, for the discussion
about hooks and stuff.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>