Initial version
It is using the master's HEAD of the git tree, but it is about the
version 0.6.0 (it's written everywhere where relevant in the code). Once
properly tagged, someday, I'll update this package to use the release
instead and increment pkgrel.

Thus why I am not adding the suffix -git to this package. Hope that's
fine.

Signed-off-by: Tomasz Bursztyka <tomasz.bursztyka@proton.me>