add support for ZeroMQ
ZeroMQ can be used in FFmpeg to dynamically control filters while
FFmpeg is running, by sending ZeroMQ messages to FFmpeg.

This is the only (good) way for scripts to interactively vary a
filter values in real time, for example to change the volume
applied to a source in real time in response to some user input.

ZeroMQ support is built in to the Debian [1] and Fedora [2]
packages at the time of writing, is quite small, and fits an
important need, so I believe it makes sense to add this to the
Archlinux build as well.

[1]: https://sources.debian.org/src/ffmpeg/7%3A7.0.2-3/debian/control/#L194
[2]: https://src.fedoraproject.org/rpms/ffmpeg/blob/main/f/ffmpeg.spec#_680