Update to version 24.9.7
I removed the Makefile patches, because I don't understand what
they're supposed to do.

Since the latest version that is downloadable as an archive from the
website is outdated, we build from a pinned GitHub repository
instead. Since the build process depends on the modification
timestamps of files, which are not preserved by Git, we have to pass
the '-W' parameter to Make.

The path '/usr/share/picolisp' has been changed to a symbolic link,
since that is what the installation instructions tell us to do.