Fixed the build process a bit.
After opening an issue on the upstream project here:

https://github.com/dev47apps/droidcam-obs-plugin/pull/25

Some suggestions were made by the upstream developer. Their current
linux/linux.mk make file allows for variables to be passed in.  I've
submitted a pull request to their project hopefully allowing for
variables representing the dynamic library names for:

* libusbmuxd
* libimobiledevice
* libturbojpeg

On Arch the current library names differ (respectively) as follows:

* libusbmuxd-2.0
* libimobiledevice-1.0
* libturbojpeg

And libturbojpeg does currently match.

fix_Makefile.patch has been renamed to fixes.patch as it currently
contains both a fix to the make file ```linux/linux.mk``` as well as
the fix (made in the previous commit) to ```src/ffmpeg_decode.cc```.

The fix for ```src/ffmpeg_decode.cc``` was mentioned in the pull request
above as a TODO so that will proably end up in the official source in
the future.  If my changes to the linux/linux.mk file are accepted into
the upstream as well this package should be reduced to just the PKGBUILD
as the fixes.patch will no longer be needed (fingers crossed).