Updated to latest commit + no patch
Removed the fixes.patch as the upstream accepted our pull request:

https://github.com/dev47apps/droidcam-obs-plugin/pull/25

While the "official" 2.3.3 tag points to commit
b931e2569ff6208eb6bb9c9c3f5723b2956cc63d (on the upstream) the latest
commit aea63f61667040480bb1a9cb12c39b74d9b98306 is the same as the
2.3.3 tagged version minus the changes to the linux/linux.mk and the
updated constant in src/ffmpeg_decode.cc.

Therefore, the _tag variable in PKGBUILD was updated to this newer
commit reducing the file count in this AUR package and bringing the
build in line with the build for droidcam-obs-plugin-git.