prepare(): do not let it strip binaries
We need symbols in order to create a debug package.  The project's
build system currently strips the symbols on its own while executing
the 'install' target, preventing makepkg to create a debug package.

Trick the build system to use 'touch -c', which does nothing, so that
makepkg can extract symbols properly and strip them (if needed).