Fix libalpm version requirements
makepkg will restrict bare library names by adding the exact version
of the library the package was built against. This behaviour does not apply
to library dependencies with explicit version requirements.

This resulted in broken 7.3.0-4 packages; the exact libalpm.so version was
not recorded in the built package.

Fix the issue by adding a separate dependency entry that makepkg will add
the appropriate version to at build time.