Switch back to make
There seems to be a bug with ninja, that I cannot quite pinpoint at the
moment, that causes a wrong dependency graph, which in turn causes a
race condition where imagemagick tries to generate files before the
dependent directory is created.

Before I go down the rabbithole to fix that bug, let's first fix the
PKGBUILD to use make as a generator again, as the generated Makefile
contains the dependency correctly.

I am not sure yet where the problem is, it could be either in ninja
itself or in the cmake generator for Ninja.