Use number of commits since start of history as version
Using tags is not suitable because the author has retroactively added tags to
older commits to denote QT5 compatibility, which means the reachability order
of the tags no longer matches their lexigraphical or chronological order.

This fact greatly confuses users and AUR helpers (in particular, `paru`) alike,
who may think that there is an update when there really isn't, or that they are
not getting the latest version.