New stable release, improved lunar phase function.
This release sets the shebang to Bash.

Updated docs.

Since v0.24:

Correctly set the correction factor of the moon phase function to "-1892" from the input year. This has to do with C practices, as C libraries commonly use year -1900. However, after extensive testing of this correction factor, it was empirically found that "-1892" works the best for this function!

Fixes errors when input month was such as "08" (which is base 10, not octal), and also improves Bash and Zsh speed on the moon phase function by almost 5x.