Switch back to files.digilent.com from direct S3 URL to support IPv6
On January 18th, 2019, the certificate for files.digilent.com expired,
and it was not replaced with a valid certificate for several weeks. This
issue caused builds of this package during that time period to fail. To
resolve the issue, I switched the PKGBUILD to use the direct URL to the
S3 bucket backing files.digilent.com, "digilent.s3.amazonaws.com".

According to certificate transparency logs, it appears that Digilent
started using Cloudflare as a caching CDN for both their main website
and files.digilent.com in January of this year. Because of this, the
certificate for files.digilent.com is now being automatically renewed
and rotated every 60 days, instead of being manually rotated once per
year as it was before.

Since we can now rely on the certificate always being up to date, and
because downloading the packages through Cloudflare's CDN enables users
on IPv6-only connections to build the package (S3 does not yet support
IPv6), let's switch back to using the files.digilent.com URL.

Fixes: 5766b4a ("Use S3 URL to work around issue with expired certificate")