Update to v1.21.42.01
This change updates the package to the upstream vendor's v1.21.42.01 of
the package. It also updates the URL of the hosted package, adds
comments to the PKGBUILD file to help with future contributions, and
most notably, it adds a new value to the DLAGENTS array in the PKGBUILD
file to avoid automated download deterrents put in place by the upstream
vendor on the download page for the source zip file.