change package to adapt to post-namechange installations
Minetest was renamed to Luanti, and this can be reflected in the
installation method, based on what package is installed. When the
`luanti-common` package is installed, then `/usr/share/minetest` is a
symlink to `/usr/share/luanti`, so the package is directly installed in
the new luanti directory instead of trying to work around the symlink.