sanitize makedeps and provides
1. as dreieck points out, having rust in the makedepends may cause
   conflicts
2. remove libvodozemac from provides()

kazv's vodozemac fork is not guaranteed to be compatible with the
original; it also has different versioning. It doesn't make sense to
provide this library by the name `libvodozemac`; The provides() is
removed unless kazv upstream renames the binary.

Technically this package should conflict libvodozemac, but currently the
original vodozemac is not packaged for arch.

Suggested-by: dreieck <oid-maps@seznam.cz>
Resolves: https://aur.archlinux.org/packages/vodozemac-bindings-cpp-kazv#comment-998251
Signed-off-by: Tianhao Wang <shrik3@mailbox.org>