do a more radical approach regarding Luanti/Minetest naming
With this update the package does a hard cut with the old name and
directories. All directories are now named after Luanti, and there are
no backwards-compatibility aliases or links anymore.

This means, all packages that depend on `minetest` from the `extra`
repository will not work any longer and need to be adjusted in order to
properly work with Luanti.

The suggested/endorsed method to install mods or games for Luanti is
using the *Content Database* anyways. (For servers, manually managing
them using Git is the common way).