rename packages to prevent future naming issues
The current Minetest packages from the extra repository might be renamed
in the future to reflect the name change to Luanti. For preventing
issues based on that name change (and to adhere to the AUR guidelines),
the packages are renamed from `luanti-...` to `luanti-modern-...`. This
also creates a better association with this split packages `pkgbase`.