upstream update: bump version to 0.2.0
also add `libvodozemac.so` to provides.

This is kinda weird, kazv vodozemac-bindings is v0.2.0 and builds
libvodozemac.so-0.2.0, while the underlying vodozemac library is 0.7.0.
Anyways I'll simply strip the version number from the provides. And
since this is the only vodozemac package let's forget about the
`conflicts` for the moment..

Signed-off-by: Tianhao Wang <shrik3@mailbox.org>