Name the downloaded git repo after the package
Makes it easier to find it in the pacman sources cache, and is more
consistent with my other packages.

NOTE: If you've previously built the package, you'll need to rename the
source file in the cache. Pacman will redownload if you don't, but that
means you have files in your cache that don't correspond to anything,
which may be confusing.
`mv "$SRCDEST/ghcid{,-static-git}`