Bump to version 1.1
Update dependency to mercury-22.01.8-2 as it includes static libraries;
otherwise setting _static_build=true at the top of the PKGBUILD fails.

Set GRADE=asm_fast.gc during the build, given that the mercury package
does not install the hlc.gc grades and the Bower build defaults to it.
Using asm_fast.gc works fine and should have similar performance.

Drop "workaround-mercury-bug567.patch" as it is included in the new
release.