fix: ensure to escape LDFLAGS when passing to external linker
The build seems to fail due to /etc/makepkg.conf setting the LDFLAGS
evironment variable to a value with spaces in it. It was introduced with
the following commit:

https://gitlab.archlinux.org/archlinux/packaging/packages/pacman/-/commit/0f10f032b02bc7823b3ac11f4807b5515c996821

The PKGBUILD for the tut package uses LDFLAGS variable to pass flags to the
external linker during the build, but does not properly escape it.

This commit fixes that, and adds myself to the list of current maintainers of
this package.