Update troubleshooting instructions
The upstream default build system is Ninja, and it does not
obey 'MAKEFLAGS'. It seems there is no environment variable[1]
to control the number of building threads with ninja. This means
the user will need to manually modify the PKGBUILD to set the
number of building threads in the cmake build command in case of
compiler error due to memory shortage.

[1] https://github.com/ninja-build/ninja/issues/1482