Re-release 1.3.0 under a new name (`sftpman` -> `sftpman-python`)
The `sftpman` name will be used by the Rust rewrite of `sftpman` (v2+).

This old Python version may still be used by some (and is a dependency
for the old Gtk-based UI frontend - `sftpman-gtk`), so it's being
republished under a new name for now. People can continue using it for a
while, if they need it.