Remove unnecessary provides/conflicts
It suffices to p/c the vanilla package since pacman considers provides
in computing conflict.

In fact, it is currently redundant to specify conflicts=(hlint-bin),
since our provides=(hlint) combines with hlint-bin's conflicts=(hlint)
to provide the conflict. Still, as noted in [1], it's best to have
provides and conflicts both just contain the vanilla pkgname, so that
other variants will definitely conflict (and there's no need for
specialized handling for any of them).

[1]: https://aur.archlinux.org/packages/hlint-bin#comment-1007225