Make conflicts more sane, use new spelling
Since pandoc (the project) has split (a while back, actually) into the
cli and other components, the vanilla package name for the dependency
has changed to suit. Use that for the dependency name.

Also, it suffices to conflict against the vanilla package, as all the
variant packages should provide=(pandoc-cli) which will make them
conflict with us given we conflicts=(pandoc-cli).