verbump: 2.0.0
Also added check()

Note: The bound on the python-bibtexparser checkdepends is due to
https://github.com/sckott/habanero/issues/144 and is provided by
aur/python-bibtexparser-git.
It can be removed once python-bibtexparser 2.0.0 lands in extra.
Alternatively, one can uncomment the `-k 'not test...'` line in the
PKGBUILD -- I've contacted upstream about this in the above issue and
hope they'll make their testsuite autodetect the presence of
`bibtexparser`.