Document incompatibility with bibtexparser 2
This was invisible until habanero (a dependency of this package) started
requiring bibtexparser>=2.0.0b5 for its tests[1].

As luck would have it, bibtexparser is merely an optdepend of habanero,
and (once [2] is merged, as it is in AUR) the tests relying on the newer
version of bibtexparser won't run if the bibtexparser version is too
low. Therefore, one can use bibtexparser<2.0.0 from extra and just
comment out that checkdepends from habanero when building it.

Raised this upstream [3] so news on this can be tracked.

[1]: https://github.com/sckott/habanero/issues/144
[2]: https://github.com/sckott/habanero/issues/199
[3]: https://github.com/sckott/habanero/issues/144