Fix build under GCC 15 (different approach)
Just pin the C version to the previous default (gnu17) instead.
The problem that I see with the previous approach, is that the BSC
packages have a lot of optional compile-time features, so it's very
tricky to ensure that I found all function declarations without
parameters, not to mention a lot of work.