Fixed install to pkgdir
This is temporary. For some reason CMake is refusing to actually install
the built files into $pkgdir.

I looked at the root CMakeLists of nuvie's source tree, and it is fairly
simple and doesn't really have any weird changes or hacks that I could
find. I'm genuinely at a loss here!

If anyone knows what the source of this issue may be and how to solve it,
feel free to hit me up on the AUR package info:

  https://aur.archlinux.org/packages/nuvie-git

Or on my email:

  rehermann6046@gmail.com