PKGBUILD: simplify how we save version of Intel's linux-clear repo
There is no point in storing the minor version of their kernel + revision
and then appending that to our major version as this might lead to errors
if one were to try to upgrade major version w/o switching to new patches.

Just save the whole version so this doesn't happen.