openvas-scanner: set BUILD_TYPE to None
Follows the guidelines for CMake based projects. As Release sets often the projects upstream Flags and overwrites in the process the ones from Arch.
This allows the removal of the workaround for the warning of redefintion the O2 with O3.
Projekt still sets -O2, but we go back to Arch default of -O3.