fix: use erlang-core instead of -nox
This also adds rebar3 to optdepends, since some gleam projects rely on
Erlang / Elixir tools to install and build dependencies.

IIRC Gleam can be used standalone to compile to Javascript, which may
not require Erlang around. If that becomes the most popular use of Gleam
in the future, maybe we could downgrade the Erlang dependencies to
optdepends.