Update to 5.3.0
The packaging changed and the webapp/ directory is gone; I think for now
it’s okay to just not copy anything else into /opt. I’m sure this could
be improved in various ways, but I’ll be honest, I haven’t actually used
this package in years; if anyone else wants to take it over, that would
probably be for the best, otherwise I’ll just try to not mess with the
existing PKGBUILD too much.