Bump to version 0.45.3
While at it, update to the latest Electron version, which works fine.
The version of Electron embedded into the Anytype packages has also
been updated (to v33), but the external versions of electron are not
detected correctly as "in production + using asar package" and the
application attempts to spawn helper programs from the wrong paths.
Therefore, add workarounds for the electron-is-dev and electron-util
Node packages.