Use a fixed-place pkgver
Use a pkgver-scheme with a fixed amount of places (4 in this case).

Because 136.0-2 released, and not 136.0.1-1 or 137.0-1, bumping to
136.0-2 of this new versioning scheme makes the old versions using the
old versioning appear newer than the new versions using the new
versioning scheme, so we need to bump epoch to 1.

Updating this package to 136.0-2 is out of scope for this branch.

This fixes the problem mentioned in the [AUR comment](https://aur.archlinux.org/packages/librewolf-bin#comment-1014287)
made by @petris.

Also credit goes to @petris on the AUR for making the base for the shell
expansions for _firefoxver and _librewolfver.