New package: ostui
ostui is a hard fork of stmps, which is itself a hard fork of stmp,
all of which are in AUR. However, ostui varies from stmps in significant
ways:

1. Browsing is by metadata, not directory structure. ostui
   shows Artists in the artists columns, not just the top-level
   directories that stmp* assumes are all artists
2. ostui is more agressive about features, such as: showing song
   metadata, album cover art, and synchronized lyrics.
3. ostui experiments with caching mechanisms to remove UI
   lag resulting from user actions hanging on network calls.
4. ostui is adding search throughout the interface, including
   search by genre, and searching for matches in any column.

Some features have been submitted upstream, and stmps includes
(e.g.) some of the search functionality. However, many
PRs are still under review and have ages in the "months",
prompting a fork.