DEV: Added a .install file and mesages to remind the admin to add users to the input group
tablet-switch requires users to be in the input group in order to toggle
hardware events. tablet-switch-git.install prints messages to facilitate
this on install and it prints a reminder on uninstall to remove users
from the input group if they're no longer required to be in it. During
the packaging step there is another reminder to add users to the input
group.