Update to version 7.1.1.47. Remove ghostpcl and ghostxps deps.
Repository ghostpcl and ghostxps now conflicts[1][2] with
ghostscript, at the same time they also need ghostscript
as a dependency.

This results in these packages not being installable anymore
(they are stil packaged in the official repositories at the
time of writing).

This is due to font licensing issues[3][4] with the official
repositories distributed packages.

References
----------

[1] https://gitlab.archlinux.org/archlinux/packaging/packages/ghostscript/-/commit/a0918031d97aa907816c5a3bdc329b18108478d8
[2] https://gitlab.archlinux.org/archlinux/packaging/packages/ghostscript/-/commit/d7d855d52ee83f73a890585f1e83ad4ad35c41ee
[3] https://gitlab.archlinux.org/archlinux/packaging/packages/ghostscript/-/issues/2
[4] https://bugs.ghostscript.com/show_bug.cgi?id=708374
[5] https://gitlab.archlinux.org/archlinux/packaging/packages/imagemagick/-/commit/b62fa42ada032c8cfbf8c471946bd7faadcc3b54