Release: 1.8.47
This changes how the package is built. We relied on GitHub releases and
tags until now, but upstream will not use these for most releases. They
prefer having a single commit to mark a new version.

This is not great, but we should still package something that make sense
for developers meanwhile.