Fix cmake compatibility
cmake4 is not compatible with cmake 3.5 or older. The upstream of lib265
already fixed this issue in the main branch, but since it's not a
release version, and the goal is to keep the same version as the
official arch linux package, the best solution is to use cmake3
temporarily.