Bumped to version 2.3.4 + Fixed package()
The previous orphaned package didn't actually package anything up and
running the build process copied files to the user's local $HOME
directory.

The package() method was completely replaced following closer to the
package() method in the droidcam-obs-plugin-git package which now copies
the plugin to /usr/lib/obs-plugins/ and the other resources to
/usr/share/obs/obs-plugins/droidcam-obs making the plugin available to
the entire system and not just the current user.

Also, !debug was added to options() since there are no debug symbols
inclued in the flatpack.zip download - no need to build a debug package.

This is named "flatpak" but it does appear to work when copied into
place.

Removed dependencies on ffmpeg and libturbojpeg as ldd revealed no
dependencies on those libraries.