Allow systemd unit to track the forked process
With `Type=oneshot`, systemd is unable to identify
or track the forked daemon process, and can’t stop
or restart it. This breaks subsequent systemd
invocations on the same unit, e.g. `systemd stop`
and `systemd restart`.

Configure the unit so that systemd correctly
guesses the main PID of the forked process, and
defers the guess until the bus name has appeared.