fix audit issues
some of the dependent build packages have a current CVE
(https://github.com/advisories/GHSA-h4j7-5rxr-p4wc)that caused
dotnet to kill the build. Instead of turning off
"WarningsAsErrors" or vuln checking, I'm just ignoring the specific CVE.
I don't believe this CVE is a threat in the case of an offline LSP like
omnisharp. it only occurs during build time, which in the case of this
package, is local to your machine. when Microsoft patches this package,
I will remove the audit supression.