Tests account for locales , fix symlink
We need to set `LC_ALL=C` to account for locales, as per
[idris-lang/Idris2#3339](https://github.com/idris-lang/Idris2/issues/3339)
(this has since been fixed, but not in a released version).

Symlinks are the bane of my existance, but the incantation should now be
correct.

With massive thanks to ElkCloner for the symlink fix, and boletales and
MithicSpirit for the locale details.