lib32-libxml2-legacy: Initial commit
Adapted from the official lib32-libxml2 and libxml2-legacy packages.
Required for capt-src, where there are a few pre-compiled 32-bit binaries that
are needed which don't have the source code, so cannot recompile against
the newer libxml2.

May be required by other packages too, but as of now, that's the only
one I know of.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>