capt-src: Add dependency to lib32-libxml2-legacy
After investigation, I found that for the most part, recompiling against
the newer libxml2 works for 95% of the binaries required for this
package to function, however, there are a few precompiled binaries in
the upstream sources, that are 32-bit, linked against libxml2.so.2, and
do not have the source code, so cannot recompile those against the newer
libxml2.

So, here's how the dependency graph looks like:

64-bit & 32-bit binaries that we compile (97% of all binaries in the end
    result package):
Linked against newer libxml2, so we require dependency `libxml2` and
`lib32-libxml2` for those.

32-bit precompiled binaries from Canon (3% of all binaries in the end
    result package):
  Linked against older libxml2, so we require dependency
  `lib32-libxml2-legacy` for these.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>