capt-src: Remove dependency to lib32-libxml2
The only binaries that depend on 32-bit libxml2 are the precompiled
binaries (which require lib32-libxml2-legacy).
we don't compile 32-bit binaries that depend on libxml2, so we don't
need dependency of lib32-libxml2.

All other binaries are 64-bit binaries that get correctly compiled and
linked against the newer version of libxml2.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>