fix: resolve openssl build failure
The package build failed to properly link to openssl. This issue took a
bit to figure out, but apparently the 'lto' setting was causing this
linker error. Thus, 'options=(!lto)' was added to resolve this.

Plus, I fixed up the way the package was constructed and its version
info. Now I am following the this format for version:

<major>.<minor>.<patch>.<revision>.<git-sha>-<build-num>

I think this is descriptive enough for this aur package!

Signed-off-by: Jason Pena <jasonpena@awkless.com>