A proper fix this time to all the bugs
Nvchad is a proper nvim plugin now, with some magic to support this.
nvchad-defaults is needed because it will get cranky if it can't find plugins
We have to add on the runtime to the top with our init.lua we cat to the starter init.lua. LAZY_NVIM_SETUP_COMPLETE being set to true stops lazy from resourcing our config a second time. I feel like this should be later in the init but it just works so maybe don't look it in the mouth?