Do not install certs to system store
The Atostek ID application's Signature Creation Service (SCS) and ERA
Smartcard ehoito features rely on self-signed certificates that can be
generated with the flags -installSCSCA and -installCA, respectively.
These certificates will be created in the user's home directory ($HOME
/.local/share/Atostek Oy/Atostek ID/). Previous versions of this AUR
package automatically generated both of these certificates and installed
them to the system's trust store. It felt like a very hacky solution,
but that's also how it's done in the upstream deb and rpm packages. In
any case, installing the certificates system wide is not enough as both
the SCS and ERA Smartcard ehoito service also require the private key,
and Atostek ID only looks for it in the users $HOME directory. Alas,
installing the certificates to the system's trust store was useless. I
have never used neither of these services, and until now I was not aware
of how they work. For this reason, it makes no sense to install the
certificates system wide anymore.