Only in KDE Plasma (Wayland) environment, the setting has been changed.
Following updates to QT and Mozc, the input position does not change from application
to application and is displayed in the top left-hand corner.
Therefore, change the setting to use classicui instead of WaylandUI as standard.

Fcitx5 displays a warning message in the Wayland environment because GTK_IM_MODULE
and QT_IM_MODULE are set, but please do not worry about it and use it as it is.
If you want to change the text or background colour, change the theme to something
like plasma in Fcitx5's classicui settings, and also change the Follow light-dark
theme setting in the system settings to ON.