Update to 3007.5 using prebuilt tarball
This commit switches from a manual relenv build to using the prebuilt
tarball. Difficulty of maintaining the manual build script as well as
significant simplification achieved by switching to the prebuilt tarball
were reason for this change.

The SALT_ADDITIONAL_PIP_PACKAGES environment variable remains
functional.

This the prebuilt tarball puts the salt executables in the onedir root
directory instead of the ./bin directory. Thus, the final destination of
the executables in the package is changed from /opt/salt/bin/ to /opt/salt/