override LDFLAGS to remove -z pack-relative-relocs
ld.gold doesn't support it. as discussed in
https://gitlab.archlinux.org/archlinux/packaging/packages/pacman/-/issues/21
ld.gold may be gone soon, in which case we may need to patch
build_scripts/qt/compilers/clang-gcc-common-switches.pri to remove the
specific dep on ld.gold which will fall back to the default of ld.bfd.
ld.lld doesn't seem to work properly with this project