fix: build from sources.
Fix PKBUILD so it builds icu from sources (based on official icu 72.1).
Only the shared libraries are provided in order to not conflict with
official icu.

The PGP signature is expired, but it should be good enough to validate
the sources.

It provides `icu72-bin` just to avoid breaking consumers. In the future
this should be removed.