Update to v0.9.0
Unfortunately I had to revoke my PGP key, but the v0.9.0 tag is still
signed with it. I don't plan to make alternative tags, but this problem
will sort itself out with the next release, that should be signed with
my new key. In the meantime, building with --skippgpcheck seems to be
the best option; there is no way to still check the signature but ignore
the revocation warning, which would still be better than not checking
anything at all.