Allow one to add daemon flags for duplicati-bin, duplicati-canary-bin, duplicati-beta-bin (#3)
* Allow one to add daemon flags for duplicati-bin, duplicati-canary-bin, duplicati-beta-bin (#1)

At time of writing, upstream duplicati allows extra flags to be defined in the systemd environment file. Add that change here. "$DAEMON_OPTS".

* Duplicati-*-bin packages, update PKGBUILD and .SRCINFO files (#2) (#3)

Forgot to update the hashes for the earlier file changes for the PKGBUILD and .SRCINFO files on duplicati-bin, duplicati-canary-bin, and duplicati-beta-bin

* Duplicati-*-bin packages, bump pkgrel version because of recent changes (#4)

Bump the pkgrel version number because of recent changes