Add a .install for host package and move the existing .install to the client and OBS plugin.
Upstream does not provide support for the Linux host component of LookingGlass.
This is explicitly mentioned in their documentation but is not read by everyone.
This .install file will explicitly tell them to not ask for support for this component.

The existing .install script was only applicable/relevant when using the client or OBS plugin.
This moves it to those packages so the user is no longer prompted to grab the Windows host component
when it is not at all relevant for those other packages.