fix checksum error
The error was with the '-xcursor' file.

The .SRCINFO file had the correct checksums but when I last commited i
did so with the PKGBUILD file having the older shasum for it, now how
did that happen you ask?

- I just found out that the pkgbuild-mode in Emacs had this option
`pkgbuild-update-sums-on-save' set to t (by default?!?)
- I never noticed it because source/asset/dist files were never
available for this function to make actual changes before
- I git added/staged the newer changes to .SRCINFO before having this
issue affect PKGBUILD, but then i guess i did something to trigger it
in PKGBUILD then commited and pushed. whoops